% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_stock_list.R
\name{complete_stock_list}
\alias{complete_stock_list}
\title{Scrapes the stock list provided by https://www.gurufocus.com/stock_list.php
in order to get ticker symbols and corresponding company names of every stock
available on gurufocus.com.}
\usage{
complete_stock_list()
}
\value{
get_stock_list() returns data.frame with the columns (char) 'Symbol'
and "Company" containing ticker symbol and full name of the public company.
}
\description{
Scrapes the stock list provided by https://www.gurufocus.com/stock_list.php
in order to get ticker symbols and corresponding company names of every stock
available on gurufocus.com.
}
