% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\docType{data}
\name{crc}
\alias{crc}
\title{CRC dataset}
\format{
\code{crc} is a list containing the following components:
\describe{
\item{otu}{the absolute abundance(count) matrix}
\item{rela}{the relative abundance matrix}
\item{batchid}{the batch(study) indicator}
\item{condition}{a factor with three levels: adenoma, CRC, control}
\item{age}{a factor with two levels: adult, senior}
\item{subtype}{a factor with five levels: adenoma, carcinoma, adenocarcinoma, advancedadenoma, control}
\item{gender}{a factor with two levels: male, female}
}
}
\usage{
crc
}
\description{
Data from curatedMetagenomicData data package
}
\keyword{datasets}
