% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gurufocus.R
\name{data_gurufocus}
\alias{data_gurufocus}
\title{Fetch complete stock list}
\usage{
data_gurufocus()
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Fetches the stock list provided by
\href{https://www.gurufocus.com/stock_list.php}{gurufocus.com/stock_list}
}
\examples{
stock_list <- data_gurufocus()

}
