% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nasdaq100.R
\name{data_nasdaq100}
\alias{data_nasdaq100}
\title{Fetch components of the Nasdaq-100}
\usage{
data_nasdaq100()
}
\value{
data.table data.frame
}
\description{
Fetches the list of components provided by
\href{https://de.wikipedia.org/wiki/NASDAQ-100}{wikipedia.org}
}
