% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sp500.R
\name{data_sp500}
\alias{data_sp500}
\title{Fetch components of the S&P 500}
\usage{
data_sp500()
}
\value{
data.table data.frame
}
\description{
Fetches the list of components provided by
\href{n.wikipedia.org/wiki/List_of_S\%26P_500_companies}{wikipedia.org}
}
