% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_turnover.R
\name{get_asset_turnover}
\alias{get_asset_turnover}
\title{Get Asset-Turnover-Ratio}
\usage{
get_asset_turnover(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Altman Z-score}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the value of a companys \emph{Net Sales Revenue}
relative to the value of its \emph{Total Assets}. The ratio indicates how
effectively a company is using its assets in order to generate sales.
}
