% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_book_value_per_share.R
\name{get_book_value_per_share}
\alias{get_book_value_per_share}
\title{Get Book Value per Share}
\usage{
get_book_value_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Book Value per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/Book+Value+Per+Share/AAPL/}{gurufocus.com}.
}
\details{
Book value per common share (or, simply book value per share - BVPS) is a
method to calculate the per-share book value of a company based on common
shareholders' equity in the company. The book value of a company is the
difference between that company's total assets and total liabilities, and not
its share price in the market.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_book_value_per_share(df)

}
