% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capex_to_sales.R
\name{get_capex_to_sales}
\alias{get_capex_to_sales}
\title{Get CAPEX-Sales-Ratio}
\usage{
get_capex_to_sales(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the companys \strong{CAPEX-Sales-Ratio} of
the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/capex_to_revenue/AAPL/}{gurufocus.com}.
}
\details{
The CAPEX-Sales-Ratio measures a companys investments into property, plant,
equipment and other capital assets (CAPEX) relative to its total sales. The
measure indicates how aggressively a company is reinvesting its revenue into
productive assets. The interpretation of the ratio depends on how effectively
a company uses its assets to produce new income.
}
