% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cash_conv_cycle.R
\name{get_cash_conv_cycle}
\alias{get_cash_conv_cycle}
\title{Get Cash Conversion Cycle}
\usage{
get_cash_conv_cycle(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Cash Conversion Cycle} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/CCC/AAPL/}{gurufocus.com}.
}
\details{
The Cash Conversion Cycle measures the amount of days it takes for a company
to convert its investments in inventory and other resources into cash flows
from sales. The measure it calculated by Days Sales Outstanding +
Days Inventory - Days Payable.

A negative value indicates that it takes a company longer to pay its
suppliers than it takes the company to sell its inventory and collect its
money. It is difficult to compare the Cash Conversion Cycle between
different industries.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_cash_conv_cycle(df)

}
