% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cash_per_share.R
\name{get_cash_per_share}
\alias{get_cash_per_share}
\title{Get Cash-per-Share}
\usage{
get_cash_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Cash-per-Share}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/cash_per_share/AAPL/}{gurufocus.com}.
}
\details{
Cash-per-Share gives the cash, cash equivalents and marketable securities
divided by the shares outstanding.
}
