% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cash_ratio.R
\name{get_cash_ratio}
\alias{get_cash_ratio}
\title{Get Cash-Ratio}
\usage{
get_cash_ratio(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Cash-Ratio} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/cash_ratio/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the value of a company's cash, cash
equivalents, marketable securities relative to its current liabilities. The
ratio indicates how liquid a company is.

It differs from the Cash-to-Debt-Ratio by focusing on current liabilities due
in the short-term (< 1 year).
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_cash_ratio(df)

}
