% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cogs_to_revenue.R
\name{get_cogs_to_revenue}
\alias{get_cogs_to_revenue}
\title{Get Cost-of-Goods-Sold-to-Revenue-Ratio}
\usage{
get_cogs_to_revenue(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Cost-of-Goods-Sold-to-Revenue-Ratio} of the last five fiscal years plus
current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/cogs2rev/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the direct cost attributed to the
production of the products sold relative to the total revenue generated by
the company over the same time period. Higher values may indicate
inefficiencies in procurement and/or production processes.
}
