% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_days_inventory.R
\name{get_days_inventory}
\alias{get_days_inventory}
\title{Get Days Inventory}
\usage{
get_days_inventory(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Days Inventory}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/DaysInventory/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio indicates the average time in day that a company
takes to turn its inventory, including goods that are work in progress,
into sales.
}
