% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_days_sales_outstanding.R
\name{get_days_sales_outstanding}
\alias{get_days_sales_outstanding}
\title{Get Days Sales Outstanding}
\usage{
get_days_sales_outstanding(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Days Sales Outstanding} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/DaysSalesOutstanding/AAPL/}{gurufocus.com}.
}
\details{
The Days Sales Outstanding is the average number of days it takes a company
to receive payment for a sale. A higher Days Sales Outstanding indicates the
company is getting its payments quickly. Generally DSO <45 days is considered
low.
}
