% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debt_to_assets.R
\name{get_debt_to_assets}
\alias{get_debt_to_assets}
\title{Get Debt to Asset Ratio}
\usage{
get_debt_to_assets(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Debt to Asset Ratio}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio is a leverage ratio that defines the total amount of
debt relative to a company's assets. A ratio of about >= 1 means a company
ownsthe same amount of liabilities or more as its assets and with that is
highly leveraged. Lower ratios indicate that a company owns more asset than
liabilities and can meet its oblications by selling assets if needed.
}
