% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debt_to_ebitda.R
\name{get_debt_to_ebitda}
\alias{get_debt_to_ebitda}
\title{Get Debt to EBITDA}
\usage{
get_debt_to_ebitda(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Debt to EBITDA} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/debt2ebitda/AAPL/}{gurufocus.com}.
}
\details{
The Debt to EBITDA is a ratio measuring the amount of income generated and
available to pay down debt before covering interest, taxes, depreciation,
and amortization expenses. Generally, net debt-to-EBITDA ratios of less than
3 are considered acceptable. The lower the ratio, the higher the probability
of the firm successfully paying off its debt. Ratios higher than 3 or 4 serve
as red flags and indicate that the company may be financially distressed in
the future.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_debt_to_ebitda(df)

}
