% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debt_to_revenue.R
\name{get_debt_to_revenue}
\alias{get_debt_to_revenue}
\title{Get Debt-to-Revenue Ratio}
\usage{
get_debt_to_revenue(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Debt to Revenue Ratio} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/debt2rev/AAPL/}{gurufocus.com}.
}
\details{
The Debt-to-Revenue Ratio is a personal finance measure that compares the
amount of debt you have to your overall income. Lenders, including issuers of
mortgages, use it as a way to measure your ability to manage the payments you
make each month and repay the money you have borrowed.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_debt_to_revenue(df)

}
