% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diluted_eps.R
\name{get_diluted_eps}
\alias{get_diluted_eps}
\title{Get Diluted Earnings per Share}
\usage{
get_diluted_eps(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
**Diluted Earnings per Share ** data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/per+share+eps/AAPL/}{gurufocus.com}.
}
\details{
The Diluted Earnings per Share calculates a company's earnings per share if
all convertible securities were converted. Dilutive securities aren't common
stock, but instead securities that can be converted to common stock.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_diluted_eps(df)

}
