% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e10.R
\name{get_e10}
\alias{get_e10}
\title{Get E10}
\usage{
get_e10(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{E10}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/e10/AAPL/}{gurufocus.com}.
}
\details{
E10 is a main component used to calculate Shiller PE Ratio. If the month end
stock price for this stock is zero, result may not be accurate due to the
exchange rate between different shares and the data will not be stored into
our database. Selected historical data showed in the calculation sectione
below is only for demostration purpose.
E10 is a concept invented by Prof. Robert Shiller, who uses E10 for his
Shiller P/E calculation. E10 is the average of the inflation adjusted
earnings of a company over the past 10 years
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_e10(df)

}
