% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebit_per_share.R
\name{get_ebit_per_share}
\alias{get_ebit_per_share}
\title{Get Earnings Before Interest and Taxes per Share}
\usage{
get_ebit_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Earnings Before Interest and Taxes per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/EBIT_per_share/AAPL/}{gurufocus.com}.
}
\details{
EBIT per Share is the amount of Earnings Before Interest and Taxes (EBIT) per
outstanding share of the company's stock.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_ebit_per_share(df)

}
