% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebitda_per_share.R
\name{get_ebitda_per_share}
\alias{get_ebitda_per_share}
\title{Get Earnings Before Interest, Tax and Depreciation Per Share}
\usage{
get_ebitda_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Get Earnings Before Interest, Tax and Depreciation Per Share} of the last
five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/EBITDA_per_share/AAPL/}{gurufocus.com}.
}
\details{
EBITDA per Share is the amount of Earnings Before Interest, Taxes,
Depreciation, and Amortization (EBITDA) per outstanding share of the
company's stock.
}
