% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps_ex_nri.R
\name{get_eps_ex_nri}
\alias{get_eps_ex_nri}
\title{Get Earnings per Share without Non-recurrent Items}
\usage{
get_eps_ex_nri(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Earnings per Share ex Non-recurrent Items} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/eps_nri/AAPL/}{gurufocus.com}.
}
\details{
The Earnings per Share without Non-recurrent Items is calculated by
subtracting non-recurring items, the dividends of preferred stocks and
non-operating income from the total net income. With that the measure gives
a better impression about the real earnings power of a company.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_eps_ex_nri(df)

}
