% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ev_to_ebit.R
\name{get_ev_to_ebit}
\alias{get_ev_to_ebit}
\title{Get Enterprise Value to Earnings before Interest Ratio}
\usage{
get_ev_to_ebit(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Enterprise Value to Earnings before Interest Ratio} of the last five
fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/ev2ebit/AAPL/}{gurufocus.com}.
}
\details{
The enterprise value to earnings before interest, taxes, depreciation, and
amortization ratio (EV/EBITDA) compares the value of a company—debt
included—to the company's cash earnings less non-cash expenses.
}
