% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ev_to_revenue.R
\name{get_ev_to_revenue}
\alias{get_ev_to_revenue}
\title{Get Enterprise-Value-to-Revenue Ratio}
\usage{
get_ev_to_revenue(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Enterprise-Value-to-Revenue Ratio} of the last five fiscal years plus
current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/ev2rev/AAPL/}{gurufocus.com}.
}
\details{
The Enterprise-Value-to-Revenue Ratio is calculated as the company's
enterprise value relative to its revenue. Often used to value a company
that does not generate income/profits yet.
}
