% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcf_per_share.R
\name{get_fcf_per_share}
\alias{get_fcf_per_share}
\title{Get Free Cash Flow per Share}
\usage{
get_fcf_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Free Cash Flow per Share} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
Free cash flow per share (FCF) is a measure of a company's financial
flexibility that is determined by dividing free cash flow by the total number
of shares outstanding. This measure serves as a proxy for measuring changes
in earnings per share.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_fcf_per_share(df)

}
