% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_financial_distress.R
\name{get_financial_distress}
\alias{get_financial_distress}
\title{Get Probability of Financial Distress}
\usage{
get_financial_distress(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's current
\strong{Probability of Financial Distress}.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/PFD/AAPL/}{gurufocus.com}.
}
\details{
The Probability of Financial Distress measures the probability that a company
will go bankrupt in the upcoming year given its current financial position.
The measure is obtained by a logit probability model based on eight
explanatory variables.
}
