% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_financial_strength.R
\name{get_financial_strength}
\alias{get_financial_strength}
\title{Get Financial Strength Rank}
\usage{
get_financial_strength(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's current years
\strong{Financial Strength Rank}.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/rank_balancesheet/AAPL/}{gurufocus.com}.
}
\details{
The Financial Strength Rank measures how strong a company's financial
situation is. The rank is base on following factors:
\itemize{
\item Interest Coverage

\item Debt-to-Revenue Ratio

\item Altman Z-score
}
Companies with a rank of 3 or less are likely to be in financial distress.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_financial_strength(df)

}
