% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_free_float.R
\name{get_free_floate}
\alias{get_free_floate}
\title{Get Float Percentage of Total Shares Outstanding}
\usage{
get_free_floate(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Float Percentage of Total Shares Outstanding} of the last five fiscal
years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/FloatPercentageOfTSO/AAPL/}{gurufocus.com}.
}
\details{
The free float percentage, also known as float percentage of total shares
outstanding, simply shows the percentage of shares outstanding that trade
freely.
}
