% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_goodwill_to_assets.R
\name{get_goodwill_to_assets}
\alias{get_goodwill_to_assets}
\title{Get Goodwill to Assets Ratio}
\usage{
get_goodwill_to_assets(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Goodwill to Asset Ratio} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/goodwill2asset/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the proportion of a company's goodwill,
which is an intangible asset, to its total assets and is a factor in that
company's valuation. The ratio quantifies a company's brand value and other
intangible aspects of its valuation.
}
