% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greenblatt_earnings_yield.R
\name{get_greenblatt_earnings_yield}
\alias{get_greenblatt_earnings_yield}
\title{Get Earnings Yield (Formula by Joel Greenblatt)}
\usage{
get_greenblatt_earnings_yield(df)
}
\arguments{
\item{df}{data.frame. Data frame with column \emph{symbol} containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Eanings Yield (Joel Greenblatt)} of the last five fiscal years as well as
it's current years TTM value.
}
\description{
Fetching \emph{Eanings Yield (Formula by Joel Greenblatt)} from gurufocus.com.
}
\details{
\emph{Joel Greenblatt's definition of Earnings Yield} has the same problems the
regular earnings yield does. It does not consider the growth of the company.
It only looks at one-year's business operation. For cyclical companies, the
earnings yield is usually highest at the peak of the business cycle, although
these earnings are rarely sustainable.
}
