% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gross_profit_to_assets.R
\name{get_gross_profit_to_assets}
\alias{get_gross_profit_to_assets}
\title{Get Gross Profit to Asset Ratio}
\usage{
get_gross_profit_to_assets(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Gross Profit to Asset Ratio} of the last five fiscal years plus current
years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio is calculated as Gross Profits divided by the firm's
Total Assets. The ratio determined how efficiently a firm uses its assets to
generate gross profits.
}
