% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insider_ownership.R
\name{get_insider_ownership}
\alias{get_insider_ownership}
\title{Get Insider Ownership}
\usage{
get_insider_ownership(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Insider Ownership}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/InsiderOwnership/AAPL/}{gurufocus.com}.
}
\details{
The Insider Ownership is the percentage of shares that are owned by company
insiders relative to the total shares outstanding. Insiders are a company's
officers, directors, relatives or generally everyone with key information
before made available to the public. High insider ownership can in many cases
be interpreted as a signal of confidence. Larger companies have typically
low(er) insider ownership.
}
