% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institutional_ownership.R
\name{get_institutional_ownership}
\alias{get_institutional_ownership}
\title{Get Institutional Ownership Percentage}
\usage{
get_institutional_ownership(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Institutional Ownership} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/InstitutionalOwnership/AAPL/}{gurufocus.com}.
}
\details{
The Institutional Ownership is the percentage of shares that are owned by
mutual or pension funds, insurance companies, investment firms, private
foundations, endowments and other large entities that manage funds on behalf
of others relative to the total shares outstanding.High institutional
ownership can in many cases be interpreted as a signal of confidence.
}
