% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interest_coverage.R
\name{get_interest_coverage}
\alias{get_interest_coverage}
\title{Get Interest Coverage}
\usage{
get_interest_coverage(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Interest Coverage} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Interest Coverage measures how easily a company ca pay interest
expenses on outstanding debt. It is calculated by diving a company's
Operating Income by its Interest Expense. Higher a coverage are naturally
better for the financial stability of a company.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_interest_coverage(df)

}
