% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory_to_revenue.R
\name{get_inventory_to_revenue}
\alias{get_inventory_to_revenue}
\title{Get Inventory to Revenue Ratio}
\usage{
get_inventory_to_revenue(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Inventory to Revenue Ratio} of the last five fiscal years plus current
years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/inventory2rev/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the percentage of inventories the company
currently has on hand to support the current amount of revenue. The ratio
indicated of a company to manage their inventory levels.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_inventory_to_revenue(df)

}
