% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory_turnover.R
\name{get_inventory_turnover}
\alias{get_inventory_turnover}
\title{Get Inventory Turnover}
\usage{
get_inventory_turnover(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Inventory Turnover} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/InventoryTurnover/AAPL/}{gurufocus.com}.
}
\details{
Inventory Turnover measures how fast the company turns over its inventory
within a year. It is calculated as Cost of Goods Sold divided by Total
Inventories.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_inventory_turnover(df)

}
