% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liabilities_to_assets.R
\name{get_liabilities_to_assets}
\alias{get_liabilities_to_assets}
\title{Get Liabilities to Assets Ratio}
\usage{
get_liabilities_to_assets(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Altman Z-score}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Liabilities to Assets Ratio is a solvency ratio indicating how much of
the company's assets are made of liabilities, calculated as total liabilities
divided by total assets. The higher the ratio is, the more risk there is in
the company.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_liabilities_to_assets(df)

}
