% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ltd_to_total_assets.R
\name{get_ltd_to_total_assets}
\alias{get_ltd_to_total_assets}
\title{Get Long-Term Debt to Total Asset Ratio}
\usage{
get_ltd_to_total_assets(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Long-Term Debt to Total Asset Ratio} of the last five fiscal years plus
current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/ltd2asset/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the percentage if a company's assets that
are financed with loans and financial obligations lasting more than one year.
The ratio gives an indication about a company's ability to meet financial
requirements for outstanding loans.
}
