% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mscore.R
\name{get_mscore}
\alias{get_mscore}
\title{Get \emph{Beneish M-Score}}
\usage{
get_mscore(df)
}
\arguments{
\item{df}{data.frame. Data frame with a column 'symbol' containing at least
one valid stock ticker symbol.}
}
\value{
The input data frame supplemented by the company's \strong{Beneish M-Score}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for \code{scrape_key_fig()} to scrape \strong{Beneish M-Score}
from \href{https://www.gurufocus.com/term/mscore/AAPL/}{gurufocus.com}.
}
\details{
The Beneish M-Score quantifies the likelihood of reported earnings manipulation.
The M-Score is a probabalistic model that classifies the rated company either
into the category "Unlikely Manipulator" (M-Score>=-1.78) or "Likely Manipulator"
(M-Score>-1.78).
The M-Score must not be applied among financial firms like banks and insurance
companies due to their very specific business characteristics.

The original formula for calculating the \strong{Beneish M-Score} consists of
eight financial ratios:
\deqn{M = -4.84+0.92xDSRI+0.528xGMI+0.404xAQI+0.892xSGI+0.115xDEPI-0.172xSGAI+
4.679xTATA-0.327xLVGI}

\itemize{
\item \eqn{X_{1}}: Ratio of working capital to total assets. Measures liquid
assets in relation to the size of the company and determines the short-term
company's solvency.

\item \eqn{X_{2}}: Ratio of retained earnings to total assets. Determines whether
the company was successful in generating profits and retaining profits for
future reinvestments in the business.

\item \eqn{X_{3}}: Ratio of earnings before interest and taxes to total assets.
Determines how effective a company is at using its own assets for generating
profits.

\item \eqn{X_{4}}: Ratio of market value of equity to book value of total liabilities.
Gives an impression to what extent the company's own assets can decline in
value before the liabilities exceed these assets and the company becomes insolvent.

\item \eqn{X_{5}}: Ratio of Sales to total assets. Common measure for determining
the total asset turnover ratio that measures how effective a company is at
generating revenue from its own assets.
}

The formula was parameterized by multivariate linear discriminant analysis applied
on a data set of 33 solvent and 33 insolvent companies (see \emph{references}).
}
\examples{
df <- data.frame('symbol' = 'AAPL')
df.altman <- get_zscore(df)

}
\references{
Beneish, Messod D. (1999): The Detection of Earnings Manipulation
in Financial Analysts Journal, Volume 55, Issue 5, p. 24-36
}
