% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_cash_per_share.R
\name{get_net_cash_per_share}
\alias{get_net_cash_per_share}
\title{Get Net Cash per Share}
\usage{
get_net_cash_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Net Cash per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/netcash/AAPL/}{gurufocus.com}.
}
\details{
Net Cash per Share is calculated by taking all a company's cash, less all
current liabilities and dividing that number by the total shares outstanding.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_net_cash_per_share(df)

}
