% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_current_asset_value.R
\name{get_net_current_asset_value}
\alias{get_net_current_asset_value}
\title{Get Net Current Asset Value}
\usage{
get_net_current_asset_value(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Net Current Asset Value} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/NCAV_real/AAPL/}{gurufocus.com}.
}
\details{
The Net Current Asset Value is a company's liquidation value. The liquidation
value is the total worth of all its physical assets, such as fixtures,
equipment, inventory and real estate. It excludes intangible assets, such as
intellectual property, brand recognition and goodwill.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_net_current_asset_value(df)

}
