% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_net_working_capital.R
\name{get_net_net_working_capital}
\alias{get_net_net_working_capital}
\title{Get Net-Net Working Capital}
\usage{
get_net_net_working_capital(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Net-Net Working Capital} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/NCAV/AAPL/}{gurufocus.com}.
}
\details{
The Net-Net Working Capital technique was developed by Benjamin Graham, in
which a company is valued based on its net-current assets per share (NCAVPS).
The Net-Net Working Capital is calculated based on current assets, taking
cash and cash equivalents at full value, then reducing accounts reseivable
for doubtful accounts and reducing inventories to liquidation values. Net-net
value is calculated by deducting total liabilities from the adjusted current
assets. Since the measure does not consider long-term assets or liabilities,
it is unreliable for long-term investments.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_net_net_working_capital(df)

}
