% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operating_cash_flow.R
\name{get_operating_cash_flow}
\alias{get_operating_cash_flow}
\title{Get Operating Cash Flow per Share}
\usage{
get_operating_cash_flow(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Operating Cash Flow per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the value of a company's
\emph{Net Sales Revenue} relative to the value of its \emph{Total Assets}. The ratio
indicates how effectively a company is using its assets in order to generate
sales.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_operating_cash_flow(df)

}
