% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owners_eps.R
\name{get_owners_eps}
\alias{get_owners_eps}
\title{Get Owners Earnings per Share}
\usage{
get_owners_eps(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Owners Earnings per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/Owner_Earnings/AAPL/}{gurufocus.com}.
}
\details{
The Owners Earnings per Share is a measure invented by Warren Buffet and
originally described in one of his famous annual shareholder letters as
follows: "If we think through these questions, we can gain some insights
about what may be called 'owner earnings.' These represent (a) reported
earnings plus (b) depreciation, depletion, amortization, and certain other
non-cash charges such as Company N's items (1) and (4) less the average
annual amount of capitalized expenditures for plant and equipment, etc. that
the business requires to fully maintain its long-term competitive position
and its unit volume. (If the business requires additional working capital to
maintain its competitive position and unit volume, the increment also should
be included in (c). However, businesses following the LIFO inventory method
usually do not require additional working capital if unit volume does not
change.)"
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_owners_eps(df)

}
