% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pe_ratio.R
\name{get_pe_ratio}
\alias{get_pe_ratio}
\title{Get Price-Earnings Ratio}
\usage{
get_pe_ratio(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Price Earnings Ratio without Non-Recurring Items} of the last five fiscal
years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/pe/AAPL/}{gurufocus.com}.
}
\details{
The Price-Earnings Ratio calculated by the earnings divided by the market
capitalization of a company.
}
