% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peg_ratio.R
\name{get_peg_ratio}
\alias{get_peg_ratio}
\title{Get Price-Earnings-Growth Ratio}
\usage{
get_peg_ratio(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Price-Earnings-Growth Ratio} of the last five fiscal years plus current
years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio is defined by the Price-to-Earnings Ratio
(without NRI) divided by the 5-Year EBITDA growth rate. A PEG Ratio >1 may
indicate overvaluation of a stock, whereas as PEG Ratio <1 may indicate
undervaluation.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_peg_ratio(df)

}
