% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piotroski_f.R
\name{get_piotroski_f}
\alias{get_piotroski_f}
\title{Get Piotroski F-Score}
\usage{
get_piotroski_f(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Piotroski F-Score} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the value of a company's
\emph{Net Sales Revenue} relative to the value of its \emph{Total Assets}. The ratio
indicates how effectively a company is using its assets in order to generate
sales.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_piotroski_f(df)

}
