% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictability.R
\name{get_predictability}
\alias{get_predictability}
\title{Get Predictability Rank}
\usage{
get_predictability(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Predictability Rank}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/predictability_rank/AAPL/}{gurufocus.com}.
}
\details{
Gurufocus.com ranks the predictability of companies based on the consistency
of their revenue per share and EBITDA (earning before interest, tax,
depreciation and amortization) per share over the past ten fiscal years, and
study the correlation between the stock performances and the predictability
of the business.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_predictability(df)

}
