% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_to_book.R
\name{get_price_to_book}
\alias{get_price_to_book}
\title{Get Price to Book Ratio}
\usage{
get_price_to_book(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Price to Book Ratio}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/pb/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the market price of a company to its book
value.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_price_to_book(df)

}
