% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_to_opcf.R
\name{get_price_to_opcf}
\alias{get_price_to_opcf}
\title{Get Price-to-Operating-Cash-Flow-Ratio}
\usage{
get_price_to_opcf(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Price-to-Operating-Cash-Flow-Ratio} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/pocf/AAPL/}{gurufocus.com}.
}
\details{
The Price-to-Operating-Cash-Flow-Ratio measures a company's value relative to
its Operating Cash Flows.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_price_to_opcf(df)

}
