% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_to_tangible_book.R
\name{get_price_to_tangible_book}
\alias{get_price_to_tangible_book}
\title{Get Price-Tangible-Book-Ratio}
\usage{
get_price_to_tangible_book(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Price-Tangible-Book-Ratio} of the last five fiscal years plus current
years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/p2tangible_book/AAPL/}{gurufocus.com}.
}
\details{
The Price-Tangible-Book-Ratio measures a company's market value to its
tangible assets. Tis ratio is applicable mainly to industrial or other
captial-intensive companies (manufacturers, miner, ...) that own a high
proportion of hard assets.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_price_to_tangible_book(df)

}
