% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profitability.R
\name{get_profitability}
\alias{get_profitability}
\title{Get Profitability Rank}
\usage{
get_profitability(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's \strong{Profitability Rank}
of the last five fiscal years plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/rank_profitability/AAPL/}{gurufocus.com}.
}
\details{
GuruFocus.com calculated the Profitability Rank by how profitable a company
is and how likely the company's business will stay that way. The maximum rank
is 10. A rank of 7 or higher means a higher profitability and may stay that
way. A rank of 3 or lower indicates that the company has had trouble to make
a profit.
The Profitability Rank is based on the Operating Margin and its 5-year
average, the Piotroski F-Score, the Consistency of the Profitability and the
Predictability Rank (see get_predicatability)
}
