% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revenue_per_share.R
\name{get_revenue_per_share}
\alias{get_revenue_per_share}
\title{Get Revenue per Share}
\usage{
get_revenue_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Revenue per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/per+share+rev/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio measures the value of a company's \emph{Revenue} relative
to its market capitalization.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_revenue_per_share(df)

}
