% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sectors.R
\name{get_sectors}
\alias{get_sectors}
\title{Get Sector and Subsector}
\usage{
get_sectors(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's sector and sub sector.
}
\description{
Function for fetching data from
\href{https://www.gurufocus.com/stock/AAPL/summary}{gurufocus.com}.
}
