% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get HTML table element}
\usage{
get_table(url, xpath, name, raw = FALSE)
}
\arguments{
\item{url}{String. The Uniform Resource Locator of the resource of interest.}

\item{xpath}{String. The XML path to an element of interest.}

\item{name}{String. The prefix preceding every scraped output column.}

\item{raw}{Logical. Controls whether 'gurufocus.com'-specific data processing
shall be applied to the table.}
}
\value{
The table to be found under the given XML path and URL.
}
\description{
Get HTML table element
}
\examples{
get_table(name = 'test',
         url = 'https://www.gurufocus.com/term/cash_per_share/AAPL/',
         xpath = '//*[@id="target_def_historical_data"]/div[2]/div/table')

}
