% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tangible_book_per_share.R
\name{get_tangible_book_per_share}
\alias{get_tangible_book_per_share}
\title{Get Tangible Book Per Share}
\usage{
get_tangible_book_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Tangible Book Per Share} of the last five fiscal years plus current years
TTM.
}
\description{
Wrapper function for fetching data from
\link{gurufocus.com}\
(https://www.gurufocus.com/term/Tangibles_book_per_share/AAPL/).
}
\details{
Tangible book value per share (TBVPS) is the value of a company's tangible
assets divided by its current outstanding shares. TBVPS determines the
potential value per share of a company in the event that it must liquidate
its assets. Assets such as property and equipment are considered tangible
assets.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_tangible_book_per_share(df)

}
