% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Get HTML text element}
\usage{
get_text(url, xpath)
}
\arguments{
\item{url}{String. The Uniform Resource Locator of the resource of interest.}

\item{xpath}{String. The XML path to an element of interest.}
}
\value{
The string to be found under the given XML path and URL.
}
\description{
Get HTML text element
}
