% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_debt_per_share.R
\name{get_total_debt_per_share}
\alias{get_total_debt_per_share}
\title{Get Total Debt Per Share}
\usage{
get_total_debt_per_share(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Total Debt per Share} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/Total_Debt_Per_Share/AAPL/}{gurufocus.com}.
}
\details{
Total Debt per Share is calculated as total debt divided by Shares
Outstanding (EOP). Total debt is calculated as Long-Term Debt & Capital Lease
Obligation plus Short-Term Debt & Capital Lease Obligation.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_total_debt_per_share(df)

}
