% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trading_volume.R
\name{get_trading_volume}
\alias{get_trading_volume}
\title{Get Trading Volume}
\usage{
get_trading_volume(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's average
\strong{Trading Volume} over the last two months.
}
\description{
Function for fetching data from
\href{https://www.gurufocus.com/stock/AAPL/summary}{gurufocus.com}.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_trading_volume(df)

}
