% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yacktman_return.R
\name{get_yacktman_forward_return}
\alias{get_yacktman_forward_return}
\title{Get Forward Rate of Return (Yacktman)}
\usage{
get_yacktman_forward_return(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Forward Rate of Return (Yacktman)} of the last five fiscal years plus
current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/RateOfReturn/AAPL/}{gurufocus.com}.
}
\details{
Yacktman defines forward rate of return as the normalized free cash flow
yield plus real growth plus inflation. 's forward rate of return for was
0.00\%. Unlike the Earnings Yield \%, the Forward Rate of Return uses the
normalized Free Cash Flow of the past seven years, and considers growth.
The forward rate of return can be thought of as the return that investors
buying the stock today can expect from it in the future.
}
