% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yoy_revenue_growth.R
\name{get_yoy_revenue_growth}
\alias{get_yoy_revenue_growth}
\title{Get Year-over-Year Revenue Per Share Growth}
\usage{
get_yoy_revenue_growth(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Year-over-Year Revenue Per Share Growth} of the last five fiscal years
plus current years TTM.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/growth_per_share_rev/AAPL/}{gurufocus.com}.
}
\details{
YoY Rev. per Sh. Growth is the percentage change of Revenue per Share over
the past twelve months
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_yoy_revenue_growth(df)

}
