% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv.R
\docType{data}
\name{hiv}
\alias{hiv}
\title{HIVRC dataset}
\format{
\code{hiv} is a list containing the following components:
\describe{
\item{otu}{the absolute abundance(count) matrix}
\item{rela}{the relative abundance matrix}
\item{groupid}{the group(study) indicator}
\item{hivstatus}{a factor with two levels}
\item{age}{a vector with continuous values}
\item{bmi}{a vector with continuous values}
\item{gender}{a factor with two levels}
}
}
\usage{
hiv
}
\description{
HIVRC dataset
}
\keyword{datasets}
