% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_data_mrmhub}
\alias{import_data_mrmhub}
\title{Import MRMhub peak integration results}
\usage{
import_data_mrmhub(data = NULL, path, import_metadata = TRUE, silent = FALSE)
}
\arguments{
\item{data}{MRMhubExperiment object}

\item{path}{One or more file paths, or a directory path (in which case all matching files will be imported)}

\item{import_metadata}{Logical, whether to import additional metadata columns (e.g., \code{batch_id}, \code{qc_type})}

\item{silent}{Logical, whether to suppress most notifications}
}
\value{
MRMhubExperiment object with the imported data
}
\description{
Imports tabular data files (*.tsv) generated from \code{MRMhub} containing peak
integration results. The input files must be in a long format with columns
for the raw data file name, feature ID, peak intensity, and other arguments
Additional information, such as retention time, FWHM, precursor/product m/z,
and CE will also be imported and made available in the \code{MRMhubExperiment}
object for downstream analyses.

When a directory path is provided, all matching files in that directory will
be imported and merged into a single dataset. This is useful when importing
datasets that were pre-processed in blocks, resulting in multiple files. Each
unique combination of feature and raw data file must only occur once across
all source data files. Duplicate combinations will result in an error.
}
\examples{
mexp <- MRMhubExperiment()

file_path = system.file("extdata", "MRMhub_demo.tsv", package = "mrmhub")

mexp <- import_data_mrmhub(
  data = mexp,
  path = file_path,
  import_metadata = TRUE)
print(mexp)
}
