% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{parse_mrmhub_result}
\alias{parse_mrmhub_result}
\title{Parses MRMhub peak integration results into a tibble}
\usage{
parse_mrmhub_result(path, silent = FALSE)
}
\arguments{
\item{path}{File name of the MRMhub result file (*.tsv or *.csv)}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Parses MRMhub peak integration results into a tibble
}
\examples{

file_path = system.file("extdata", "MRMhub_demo.tsv", package = "mrmhub")

tbl <- parse_mrmhub_result(path = file_path)

head(tbl)
}
