% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuanT-package.R
\docType{package}
\name{QuanT-package}
\alias{QuanT-package}
\title{QuanT: QuanT}
\description{
Identifying unmeasured heterogeneity in microbiome data via quantile thresholding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jiuyaolu/QuanT}
  \item Report bugs at \url{https://github.com/jiuyaolu/QuanT/issues}
}

}
\author{
\strong{Maintainer}: Jiuyao Lu \email{lujiuyao@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
