% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\docType{package}
\name{quant}
\alias{quant}
\alias{_PACKAGE}
\alias{quant-package}
\title{flightsbr: Download Flight And Airport Data from Brazil}
\description{
Download flight and airport data from Brazil’s Civil Aviation Agency (ANAC)
\url{https://www.gov.br/anac}. The data includes detailed information on all
aircrafts, aerodromes, airports, and airports movements registered in ANAC,
and on every international flight to and from Brazil, as well as domestic
flights within the country.
}
\section{Usage}{

Please check the vignettes on the \href{https://ipeagit.github.io/flightsbr/}{website}.
}

\author{
\strong{Maintainer}: Oliver Hennhöfer \email{oliver.hennhoefer@mail.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
