% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_ttm.R
\name{regex_ttm}
\alias{regex_ttm}
\title{Text-processing of key figure from gurufocus.com}
\usage{
regex_ttm(string)
}
\arguments{
\item{string}{String. The character string to be processed}
}
\value{
The input data frame supplemented by the company's respective key figures
of the last five fiscal years plus current years TTM.
}
\description{
Internal \strong{'quant'}-function for processing the scraped character string
of key figures provided by gurufocus.com
}
