% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{sanity}
\alias{sanity}
\title{Apply sanity checks}
\usage{
sanity(df)
}
\arguments{
\item{df}{Data.frame. A data.frame to be checked for sanity.}
}
\description{
Internal function. Applies a sanity check on given input parameter(s)
}
\examples{
sanity(data.frame("symbol" = "AAPL", "MSFT", "BABA)) #passes
sanity(data.frame("stocks" = "AAPL", "MSFT", "BABA)) #flunks
}
