% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ownership_fig.R
\name{scrape_ownership_fig}
\alias{scrape_ownership_fig}
\title{Get Key Figure from gurufocus.com}
\usage{
scrape_ownership_fig(df, url, pfx, xpath_txt, xpath_tbl)
}
\arguments{
\item{df}{data.frame. Data frame with a column 'symbol' containing at least
one valid stock ticker symbol.}

\item{url}{String. The Uniform Resource Locator of the resource of interest.}

\item{pfx}{String. Prefix for column names of scraped data sets.}

\item{xpath_txt}{String. The XML path to a text element of interest.}

\item{xpath_tbl}{String. The XML path to a table element of interest.}
}
\value{
The input data frame supplemented by a company's respective key figure value
of the last five fiscal years plus current years TTM.
}
\description{
Function for fetching data from
\href{https://www.gurufocus.com/term/InstitutionalOwnership/AAPL/}{gurufocus.com}.
}
