% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_predictability.R
\name{scrape_predictability}
\alias{scrape_predictability}
\title{Helper Function for Predictability}
\usage{
scrape_predictability(df, url, xpath_txt)
}
\arguments{
\item{df}{data.frame}

\item{url}{String}

\item{xpath_txt}{String}
}
\value{
data.frame
}
\description{
Helper Function for Predictability
}
