% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_profitability.R
\name{scrape_profitability}
\alias{scrape_profitability}
\title{Helper Function for Profitability

@usage scrape_profitability(df, url, xpath_txt)}
\usage{
scrape_profitability(df, url, xpath_txt)
}
\arguments{
\item{df}{data.frame}

\item{url}{String}

\item{xpath_txt}{String}
}
\value{
data.frame
}
\description{
Helper Function for Profitability

@usage scrape_profitability(df, url, xpath_txt)
}
