% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicality.R
\name{topicality}
\alias{topicality}
\title{Merge original data frame with scraped data

Internal function. Merges data.frame resulting from web scraping to the
data.frame provided as input of a respective function.}
\usage{
topicality(table)
}
\arguments{
\item{table}{Data.frame}
}
\description{
Merge original data frame with scraped data

Internal function. Merges data.frame resulting from web scraping to the
data.frame provided as input of a respective function.
}
\examples{
df <- data.frame("symbol" = "STOCK", "eps_2016" = 99)
topicality(df)
}
