% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nasdaq100.R
\name{data_nasdaq100}
\alias{data_nasdaq100}
\title{Fetch quotes for NASDAQ-100 Index}
\usage{
data_nasdaq100()
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Fetches the stock list provided by
\href{https://de.wikipedia.org/wiki/NASDAQ-100}{wikipedia.org}
}
\examples{
nasdaq100 <- data_nasdaq100()

}
