% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sp500.R
\name{data_sp500}
\alias{data_sp500}
\title{Fetch quotes for S&P500 Index}
\usage{
data_sp500()
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Fetches the stock list provided by
\href{n.wikipedia.org/wiki/List_of_S\%26P_500_companies}{wikipedia.org}
}
\examples{
sp500 <- data_sp500()

}
