% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_ratio.R
\name{get_current_ratio}
\alias{get_current_ratio}
\title{Get Current Ratio}
\usage{
get_current_ratio(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Current Ratio} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/current_ratio/AAPL/}{gurufocus.com}.
}
\details{
The Current Ratio measures a company's ability to its shot-term obligations.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_current_ratio(df)

}
