% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eff_interest_rate.R
\name{get_eff_interest_rate}
\alias{get_eff_interest_rate}
\title{Effective Annual Interest Rate}
\usage{
get_eff_interest_rate(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Effective Annual Interest Rate} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/turnover/AAPL/}{gurufocus.com}.
}
\details{
The Effective Annual Interest Rate is the interest rate on a loan restated
from the nominal interest rate and expressed as if compound interest was
payable annually. It makes interest rates between loans with different
compounding periods more comparable.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_eff_interest_rate(df)

}
