% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greenblatt_earnings_yield.R
\name{get_greenblatt_earnings_yield}
\alias{get_greenblatt_earnings_yield}
\title{Get Eanings Yield (Joel Greenblatt)}
\usage{
get_greenblatt_earnings_yield(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
The original data.frame supplemented by the company's
\strong{Eanings Yield (Joel Greenblatt)} of the last five fiscal years plus
current years TTM.
}
\description{
Wrapper function for fetching data from
\link{gurufocus.com}\
(https://www.gurufocus.com/term/earning_yield_greenblatt/AAPL/).
}
\details{
Joel Greenblatt's definition of earnings yield has the same problems the
regular earnings yield does. It does not consider the growth of the company.
It only looks at one-year's business operation. For cyclical companies, the
earnings yield is usually highest at the peak of the business cycle. But
these earnings are rarely sustainable.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_greenblatt_earnings_yield(df)

}
