% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pe_ratio_nri.R
\name{get_pe_ratio_nri}
\alias{get_pe_ratio_nri}
\title{Get Price-Earnings Ratio without Non-Recurring Items}
\usage{
get_pe_ratio_nri(df)
}
\arguments{
\item{df}{data.frame. Data frame with column 'symbol' containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Price Earnings Ratio with Non-Recurring Items} data.
}
\description{
Wrapper function for fetching data from
\href{https://www.gurufocus.com/term/penri/AAPL/}{gurufocus.com}.
}
\details{
The Asset-Turnover-Ratio gives the price to earnings ratio without the
potentially misleading effects if non-recurring items (e.g. sale of a major
asset). The ratio only considers regular operating income.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_pe_ratio_nri(df)

}
