% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tangible_book_per_share.R
\name{get_tangible_book_per_share}
\alias{get_tangible_book_per_share}
\title{Get Tangible Book Value Per Share}
\usage{
get_tangible_book_per_share(df)
}
\arguments{
\item{df}{data.frame. Data.frame with column \emph{symbol} containing at least one
valid stock ticker symbol.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Tangible Book Per Share} data.
}
\description{
Fetching \emph{Tangible Book Value Per Share} from gurufocus.com
}
\details{
The \emph{Tangible book value per share} is the value of a company's tangible
assets divided by its current outstanding shares. The TBVPS determines the
potential value per share of a company in the event that it must liquidate
it's assets.
Assets such as property and equipment are considered tangible assets.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_tangible_book_per_share(df)

}
