% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yoy_ebitda_growth.R
\name{get_yoy_ebitda_growth}
\alias{get_yoy_ebitda_growth}
\title{Get Year-over-Year Earnings before Interest, Tax and
Appreciation Growth-Rate}
\usage{
get_yoy_ebitda_growth(df)
}
\arguments{
\item{df}{data.frame. Data.frame with column \emph{symbol} containing at least one
valid ticker symbol of a listed stock.}
}
\value{
Input data.frame supplemented by the company's available
\strong{Year-over-Year Earnings before Interest, Tax and Appreciation Growth}
data.
}
\description{
Fetching
\emph{Year-over-Year Earnings before Interest, Tax and Appreciation Growth-Rate}
from gurufocus.com
}
\details{
\emph{YoY EBITDA Growth} is the percentage change of EBITDA per share.
}
\examples{
df <- data.frame('symbol' = 'AAPL')
res <- get_yoy_ebitda_growth(df)

}
