% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmhub-package.R
\docType{package}
\name{mrmhub-package}
\alias{mrmhub}
\alias{mrmhub-package}
\title{mrmhub: Small Molecule Mass Spectrometry Data Post-Processing, Quality Control, Exploration and Reporting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'mrmhub' package offers R functions for processing, quality control, and visualization of lipidomics mass spectrometry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SLINGhub/mrmhub}
  \item \url{https://slinghub.github.io/mrmhub/}
  \item \url{https://slinghub.github.io/MRMhub/}
  \item Report bugs at \url{https://github.com/SLINGhub/mrmhub/issues}
}

}
\author{
\strong{Maintainer}: Bo Burla \email{bo.burla@nus.edu.sg} (\href{https://orcid.org/0000-0002-5918-3249}{ORCID})

}
\keyword{internal}
