% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuanT.R
\name{QuanT}
\alias{QuanT}
\title{QuanT}
\usage{
QuanT(dat, covariates, q = NULL)
}
\arguments{
\item{dat}{The abundance matrix with rows as samples and columns as taxa.
This is supposed to be relative abundances after preprocessing.}

\item{covariates}{All the known variables including the primary variable of
interest and other covariates.}

\item{q}{The quantile levels to be used. If NULL, QuanT will choose the
quantile levels based on the proportion of zeroes in the data.}
}
\value{
\code{QuanT} outputs the QSVs, which are surrogates for the
unmeasured heterogeneity in the data. They are stored in a matrix with the
same number of rows as \code{dat}.
}
\description{
\code{QuanT} runs the QuanT algorithm. It takes as input the abundance
matrix and covariates, and outputs the QSVs.
}
\examples{
covariates = model.matrix(~ subtype + age + gender, data=crc)[,-1]
qsv = QuanT(crc$rela, covariates)
plot(qsv[,1:2], col=crc$batchid)
}
